can: mcp25xxfd: mcp25xxfd_can_queue_frame(): change argument "ts" to u32
The timestamp is an u32, so don't cap to upper 16 bit.

(cherry picked from commit 6546a196615cae17d1e4c52f5135b02469adc15f)

+++ drivers/net/can/spi/mcp25xxfd/mcp25xxfd_can.h
@@ -23,7 +23,7 @@ int mcp25xxfd_can_targetmode(struct mcp25xxfd_can_priv *cpriv)
 
 static inline
 void mcp25xxfd_can_queue_frame(struct mcp25xxfd_can_priv *cpriv,
-			       s32 fifo, u16 ts, bool is_rx)
+			       s32 fifo, u32 ts, bool is_rx)
 {
 	int idx = cpriv->fifos.submit_queue_count;
